Added m_joinpartsno for a join/part snomask
